AdaMLearn/requirements.txt
backpack_for_pytorch==1.6.0
matplotlib==3.5.2
numpy==2.4.0
pandas==1.4.4
Pillow==9.2.0
Pillow==12.1.0
prefetch_generator==1.0.1
Requests==2.32.5
scikit_learn==1.0.2
scipy==1.16.3
seaborn==0.13.2
six==1.16.0
torch==2.5.1
torchvision==0.20.1
